// dashboard/src/api/models/noteVisibilityEnum.ts
/**
 * Generated by orval v7.17.0 🍺
 * Do not edit manually.
 * Invoicence API
 * Description
 * OpenAPI spec version: 1.0.0
 */

/**
 * * `internal` - Internal
 * `public` - Public
 */
export type NoteVisibilityEnum =
  (typeof NoteVisibilityEnum)[keyof typeof NoteVisibilityEnum];

// eslint-disable-next-line @typescript-eslint/no-redeclare
export const NoteVisibilityEnum = {
  internal: "internal",
  public: "public",
} as const;