// dashboard/src/api/models/note.ts
/**
 * Generated by orval v7.17.0 🍺
 * Do not edit manually.
 * Invoicence API
 * Description
 * OpenAPI spec version: 1.0.0
 */
import type { NoteAuthor } from "./noteAuthor";
import type { NoteVisibilityEnum } from "./noteVisibilityEnum";

export interface Note {
  id: string;
  /** @maxLength 2048 */
  content: string;
  visibility: NoteVisibilityEnum;
  created_at: string;
  author: NoteAuthor;
}