// dashboard/src/api/models/noteCreate.ts
/**
 * Generated by orval v7.17.0 🍺
 * Do not edit manually.
 * Invoicence API
 * Description
 * OpenAPI spec version: 1.0.0
 */
import type { NoteVisibilityEnum } from "./noteVisibilityEnum";

export interface NoteCreate {
  /** @maxLength 2048 */
  content: string;
  visibility: NoteVisibilityEnum;
}